+++ tools/test-js-with-node
@@ -67,6 +67,7 @@ EXEMPT_FILES = make_set(
         "web/src/compose_fade.js",
         "web/src/compose_recipient.js",
         "web/src/compose_state.js",
+        "web/src/compose_textarea.js",
         "web/src/compose_ui.js",
         "web/src/compose_validate.js",
         "web/src/composebox_typeahead.js",
